server/requirements.txt
Flask
Flask-Cors
requests
psycopg2-binary
bcrypt
